nit refactor
